ts-model/requirements.txt
scikit-learn
websocket-client
pandas
numpy
pyspark
joblib
requests
fastapi
uvicorn
apscheduler
